requirements.txt
anyio==4.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.9.23
fonttools==4.61.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.6
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.45.1
MarkupSafe==2.1.5
matplotlib==3.10.7
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.3.5
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnx==1.17.0
onnxruntime-gpu==1.22.0
packaging==25.0
pillow==12.0.0
protobuf==6.33.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.3
setuptools==70.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
tifffile==2025.10.16
torch==2.9.1+cu128
tqdm==4.67.1
triton==3.5.1
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.5.0
